extendedglmnet: an extension to glmnet package with options of lasso, ridge and random lasso methods

The extendedglmnet package provides four functions: extendedglmnet, print.extendglmnet, predict.extendglmnet, and evaluation.

This package is an extension of glmnet package. Users can either fit a linear model or generalized linear model depending the response variable types. It can support ridge, lasso and random lasso methods. User can choose to use cross validation to find the optimal regularization parameter lambda or use user specified lambda. Can deal with both continuous and binary responses.

extendedglmnet(
  x,
  y,
  family = c("gaussian", "binomial"),
  type = c("regression", "ridge", "lasso", "random lasso"),
  cv = TRUE,
  lambda = 0,
  nfolds = 5,
  B = 200,
  q1 = 5,
  q2 = 5
)

Arguments

x

input matrix, of dimension nobs x nvars; each row is an observation vector.

y

y response variable. For family="gaussian", y can be any real values. For family="binomial" should be a factor with two levels.

family

a character string representing one of the built-in families gaussian or binomial.

type

a character string representing one of the built-in type of regression methods regression, ridge, lasso or random lasso.

cv

an indicator variable to specify whether cross validation should be used to pick optimal lambda value. Default is TRUE.

lambda

A user supplied lambda sequence. Typical usage is to have the program compute its own lambda sequence. Supplying a value of lambda overrides this.

nfolds

number of folds - default is 5. Although nfolds can be as large as the sample size (leave-one-out CV), it is not recommended for large datasets. Smallest value allowable is nfolds=3

B

number of bootstrap samples in random lasso. One can take B = 500 or B = 1000. The default is 200.

q1

number of candidate variables to be included in each bootstrap sample during generating importance measures in random lasso. This value need to be tuned to ensure best performance. The value should be positive and no greater than number of predictors in x.

q2

number of candidate variables to be included in each bootstrap sample during variable selection in random lasso. This value need to be tuned to ensure best performance. The value should be positive and no greater than number of predictors in x.

Value

a list containing fitting results.

type

the type of regression performed based upon user specified type.

family

the regression family based upon response variable type.

coef

a list of estimated coefficient.

lambda

the actual sequence of lambda values used. It is always 0 for regression models without regulation.

df

number of nonzero coefficients for each value of lambda.

nobs

number of observations.

dim

dimension of coefficient matrix.

Details

Lasso is known to be suffering from that it can select only one or a few of a set of highly correlated important variables. If several independent data sets were generated from the same distribution, then we would expect lasso to select nonidentical subsets of those highly correlated important variables from different data sets. Random Lasso proposed by Sijian Wang et al. solved this issue by using bootstrap, which will provide more stable results than lasso and ridge when number of predictors are greater than number of observations.

References

Wang, S., Nan, B., Rosset, S., & Zhu, J. (2011). Random lasso, The annals of applied statistics, 5(1), 468.
Friedman J, Hastie T, Tibshirani R (2010). Regularization Paths for Generalized Linear Models via Coordinate Descent. Journal of Statistical Software, 33(1), 1–22. doi: 10.18637/jss.v033.i01, https://www.jstatsoft.org/v33/i01/.

See also

Author

Weiwei Tao tww101928@gmail.com

Examples

# a binary response example
size = 200
x = matrix(rnorm(210*size),ncol=210)
z = 1 + 2*x[,1] + 3*x[,40] - 4*x[,50] + 3*x[,210]
pr = 1/(1+exp(-z))
y = rbinom(size,1,pr)

index = sample(1:size, 0.7*size)
trainx = x[index,]
testx = x[-index,]
trainy = y[index]
testy = y[-index]

# logistic regression
fit.glm <- extendedglmnet(trainx, trainy, family = "binomial", type = "regression")
#> Warning: glm.fit: algorithm did not converge
print.extendglmnet(fit.glm)
#> 
#> Call: logistic regression for binomial reponses
#> 
#>  (Intercept)           x1           x2           x3           x4           x5 
#>  -50.5256476   14.0172243   12.2439779  -89.7125513  -18.0241232   -1.5476567 
#>           x6           x7           x8           x9          x10          x11 
#>  -29.4867663   47.1960716   38.0826959   26.9512104  -49.4493838   65.0692003 
#>          x12          x13          x14          x15          x16          x17 
#>  -41.8155752   73.7033428   45.0835312   33.1742724   21.2367490   16.7907859 
#>          x18          x19          x20          x21          x22          x23 
#>   53.6864330    1.7400215   44.4504166  -66.6390614   64.5606759   38.7179860 
#>          x24          x25          x26          x27          x28          x29 
#>    9.9141729   14.0570997   11.8134811   -7.4308068   28.6646494  -20.9336874 
#>          x30          x31          x32          x33          x34          x35 
#>    4.6168767   17.4287423  -56.2077676  -27.4116795   24.0736220   35.6030495 
#>          x36          x37          x38          x39          x40          x41 
#>  -47.7231698  -18.2144578  -16.2628726  -87.6010518  -18.5685311  -13.8779148 
#>          x42          x43          x44          x45          x46          x47 
#>  138.8273118  -93.8135237  -52.9706209   -2.1969681   66.7499117  -33.7947980 
#>          x48          x49          x50          x51          x52          x53 
#>  -68.3271098  -75.3490113  -30.5777584   -2.0322207  -85.8189037   16.7996887 
#>          x54          x55          x56          x57          x58          x59 
#>   -9.3772111  -14.4061598   48.2038093    8.6449532   -4.6573402   -6.3271335 
#>          x60          x61          x62          x63          x64          x65 
#>   79.0157934   41.4977537   34.4968214  -33.5667286  -27.8586657    4.9373862 
#>          x66          x67          x68          x69          x70          x71 
#>  -18.9377807    5.8244495   54.9530546   15.5589081   -0.6413093   14.6117713 
#>          x72          x73          x74          x75          x76          x77 
#>   34.3521300  -19.3728142    3.7452735  -17.5036021   -0.7910131    8.6552518 
#>          x78          x79          x80          x81          x82          x83 
#>   79.5381114  -25.0714178  -15.2672571  -39.5760956  -41.0731514  -14.0027090 
#>          x84          x85          x86          x87          x88          x89 
#>   87.6086334  -26.5309332    7.3762491  -50.7510536   18.3852491   53.7322604 
#>          x90          x91          x92          x93          x94          x95 
#> -102.1727753    4.7288271  -19.8521012   30.5488323   75.6429656  -32.9153954 
#>          x96          x97          x98          x99         x100         x101 
#>  -50.4716323  -42.5136452   29.2693311  -77.2496089   -9.4579732  -58.4817931 
#>         x102         x103         x104         x105         x106         x107 
#>   11.5084811 -102.7466727  -13.1563479   31.8281730   19.8916434  -41.4407568 
#>         x108         x109         x110         x111         x112         x113 
#>  -54.5943520   -7.7890152   20.1384688   35.5927263   37.7624357   38.1803085 
#>         x114         x115         x116         x117         x118         x119 
#>    7.1230792   13.6634459   11.7814726  -15.1416325   26.5614049  -38.4384280 
#>         x120         x121         x122         x123         x124         x125 
#>   -6.1374046   89.7927001  -34.3158032  -64.7215648  -84.9952205   35.6782811 
#>         x126         x127         x128         x129         x130         x131 
#>   42.7891818   29.6259335   31.4770857  -45.3581605    6.9791271   -5.1053102 
#>         x132         x133         x134         x135         x136         x137 
#>   -7.2332280  -11.7974376  -29.9634387  -11.5346316  -43.5332796   51.5308738 
#>         x138         x139         x140         x141         x142         x143 
#>    3.8826226    8.2577446           NA           NA           NA           NA 
#>         x144         x145         x146         x147         x148         x149 
#>           NA           NA           NA           NA           NA           NA 
#>         x150         x151         x152         x153         x154         x155 
#>           NA           NA           NA           NA           NA           NA 
#>         x156         x157         x158         x159         x160         x161 
#>           NA           NA           NA           NA           NA           NA 
#>         x162         x163         x164         x165         x166         x167 
#>           NA           NA           NA           NA           NA           NA 
#>         x168         x169         x170         x171         x172         x173 
#>           NA           NA           NA           NA           NA           NA 
#>         x174         x175         x176         x177         x178         x179 
#>           NA           NA           NA           NA           NA           NA 
#>         x180         x181         x182         x183         x184         x185 
#>           NA           NA           NA           NA           NA           NA 
#>         x186         x187         x188         x189         x190         x191 
#>           NA           NA           NA           NA           NA           NA 
#>         x192         x193         x194         x195         x196         x197 
#>           NA           NA           NA           NA           NA           NA 
#>         x198         x199         x200         x201         x202         x203 
#>           NA           NA           NA           NA           NA           NA 
#>         x204         x205         x206         x207         x208         x209 
#>           NA           NA           NA           NA           NA           NA 
#>         x210 
#>           NA 
#> Degrees of Freedom: 139 Total (i.e. Null); 0 Residual
#> 
ypred <- predict.extendglmnet(fit.glm, testx, family = "binomial", type="class")
#> Warning: prediction from a rank-deficient fit may be misleading
evaluation(testy, ypred, "accuracy")
#> [1] 0.4666667

# lasso
fit.lasso <- extendedglmnet(trainx, trainy, family = "binomial", type = "lasso")
#> <sparse>[ <logic> ] : .M.sub.i.logical() maybe inefficient
print.extendglmnet(fit.lasso)
#> 
#> Call: lasso for binomial reponses
#> 
#> 211 x 1 sparse Matrix of class "dgCMatrix"
#>                       s0
#> (Intercept)  0.334037369
#> x1           0.311977966
#> x2           .          
#> x3           .          
#> x4           .          
#> x5           .          
#> x6           .          
#> x7          -0.080133228
#> x8           .          
#> x9           .          
#> x10          .          
#> x11          .          
#> x12          .          
#> x13          .          
#> x14          .          
#> x15          0.229489601
#> x16          .          
#> x17          .          
#> x18          .          
#> x19          .          
#> x20          .          
#> x21          .          
#> x22          .          
#> x23          .          
#> x24          .          
#> x25          .          
#> x26          .          
#> x27          .          
#> x28         -0.096552340
#> x29          .          
#> x30          .          
#> x31          .          
#> x32          .          
#> x33          .          
#> x34          .          
#> x35          .          
#> x36          .          
#> x37          .          
#> x38          .          
#> x39          .          
#> x40          0.956241108
#> x41          .          
#> x42          .          
#> x43          .          
#> x44          .          
#> x45          .          
#> x46          .          
#> x47          .          
#> x48          .          
#> x49          .          
#> x50         -1.425708042
#> x51          .          
#> x52          .          
#> x53         -0.056173974
#> x54          .          
#> x55         -0.124751844
#> x56          0.080063331
#> x57          .          
#> x58          .          
#> x59          .          
#> x60          .          
#> x61          .          
#> x62          .          
#> x63          0.050205729
#> x64          .          
#> x65          .          
#> x66          .          
#> x67          .          
#> x68          .          
#> x69          .          
#> x70          .          
#> x71          .          
#> x72          .          
#> x73          .          
#> x74          .          
#> x75          .          
#> x76          .          
#> x77          .          
#> x78          .          
#> x79         -0.007024797
#> x80         -0.011830803
#> x81          .          
#> x82          .          
#> x83          .          
#> x84          .          
#> x85          .          
#> x86          .          
#> x87         -0.050411155
#> x88          .          
#> x89          .          
#> x90          .          
#> x91          .          
#> x92          .          
#> x93          0.124898472
#> x94          .          
#> x95          .          
#> x96          .          
#> x97          .          
#> x98          .          
#> x99          .          
#> x100        -0.032992885
#> x101         0.223830451
#> x102         .          
#> x103         .          
#> x104         0.073964984
#> x105         .          
#> x106         .          
#> x107         .          
#> x108         .          
#> x109         .          
#> x110         .          
#> x111         .          
#> x112         .          
#> x113         .          
#> x114         .          
#> x115         .          
#> x116         0.041208418
#> x117         .          
#> x118         .          
#> x119         .          
#> x120         .          
#> x121         .          
#> x122        -0.114895998
#> x123         .          
#> x124         .          
#> x125         .          
#> x126         .          
#> x127         .          
#> x128         .          
#> x129         .          
#> x130         .          
#> x131         .          
#> x132         .          
#> x133         .          
#> x134         .          
#> x135         0.050432555
#> x136         .          
#> x137         .          
#> x138         .          
#> x139         .          
#> x140         .          
#> x141         .          
#> x142         .          
#> x143         .          
#> x144         .          
#> x145         .          
#> x146         .          
#> x147         .          
#> x148         .          
#> x149         .          
#> x150         .          
#> x151        -0.058158645
#> x152         .          
#> x153         .          
#> x154         .          
#> x155         .          
#> x156         .          
#> x157         .          
#> x158         .          
#> x159         0.089086626
#> x160         .          
#> x161         .          
#> x162         .          
#> x163         .          
#> x164         .          
#> x165         .          
#> x166         .          
#> x167         .          
#> x168        -0.288060753
#> x169         .          
#> x170         .          
#> x171         .          
#> x172         .          
#> x173         .          
#> x174         .          
#> x175         .          
#> x176         .          
#> x177         .          
#> x178         .          
#> x179         .          
#> x180         .          
#> x181         0.037796264
#> x182         .          
#> x183         0.042028658
#> x184         .          
#> x185         .          
#> x186         .          
#> x187         .          
#> x188         .          
#> x189         .          
#> x190         .          
#> x191         .          
#> x192         .          
#> x193         .          
#> x194         .          
#> x195         .          
#> x196         .          
#> x197         .          
#> x198         .          
#> x199         .          
#> x200         .          
#> x201         .          
#> x202        -0.019157694
#> x203         .          
#> x204         .          
#> x205         .          
#> x206         .          
#> x207         .          
#> x208         .          
#> x209         .          
#> x210         0.730390941
#> Degrees of Freedom: 139 Total (i.e. Null); 112 Residual
#> 
ypred <- predict.extendglmnet(fit.lasso, testx, family = "binomial", type="class")
evaluation(testy, ypred, "accuracy")
#> [1] 0.8833333

# ridge
fit.ridge <- extendedglmnet(trainx, trainy, family = "binomial", type = "lasso")
#> <sparse>[ <logic> ] : .M.sub.i.logical() maybe inefficient
print.extendglmnet(fit.ridge)
#> 
#> Call: lasso for binomial reponses
#> 
#> 211 x 1 sparse Matrix of class "dgCMatrix"
#>                       s0
#> (Intercept)  0.334037369
#> x1           0.311977966
#> x2           .          
#> x3           .          
#> x4           .          
#> x5           .          
#> x6           .          
#> x7          -0.080133228
#> x8           .          
#> x9           .          
#> x10          .          
#> x11          .          
#> x12          .          
#> x13          .          
#> x14          .          
#> x15          0.229489601
#> x16          .          
#> x17          .          
#> x18          .          
#> x19          .          
#> x20          .          
#> x21          .          
#> x22          .          
#> x23          .          
#> x24          .          
#> x25          .          
#> x26          .          
#> x27          .          
#> x28         -0.096552340
#> x29          .          
#> x30          .          
#> x31          .          
#> x32          .          
#> x33          .          
#> x34          .          
#> x35          .          
#> x36          .          
#> x37          .          
#> x38          .          
#> x39          .          
#> x40          0.956241108
#> x41          .          
#> x42          .          
#> x43          .          
#> x44          .          
#> x45          .          
#> x46          .          
#> x47          .          
#> x48          .          
#> x49          .          
#> x50         -1.425708042
#> x51          .          
#> x52          .          
#> x53         -0.056173974
#> x54          .          
#> x55         -0.124751844
#> x56          0.080063331
#> x57          .          
#> x58          .          
#> x59          .          
#> x60          .          
#> x61          .          
#> x62          .          
#> x63          0.050205729
#> x64          .          
#> x65          .          
#> x66          .          
#> x67          .          
#> x68          .          
#> x69          .          
#> x70          .          
#> x71          .          
#> x72          .          
#> x73          .          
#> x74          .          
#> x75          .          
#> x76          .          
#> x77          .          
#> x78          .          
#> x79         -0.007024797
#> x80         -0.011830803
#> x81          .          
#> x82          .          
#> x83          .          
#> x84          .          
#> x85          .          
#> x86          .          
#> x87         -0.050411155
#> x88          .          
#> x89          .          
#> x90          .          
#> x91          .          
#> x92          .          
#> x93          0.124898472
#> x94          .          
#> x95          .          
#> x96          .          
#> x97          .          
#> x98          .          
#> x99          .          
#> x100        -0.032992885
#> x101         0.223830451
#> x102         .          
#> x103         .          
#> x104         0.073964984
#> x105         .          
#> x106         .          
#> x107         .          
#> x108         .          
#> x109         .          
#> x110         .          
#> x111         .          
#> x112         .          
#> x113         .          
#> x114         .          
#> x115         .          
#> x116         0.041208418
#> x117         .          
#> x118         .          
#> x119         .          
#> x120         .          
#> x121         .          
#> x122        -0.114895998
#> x123         .          
#> x124         .          
#> x125         .          
#> x126         .          
#> x127         .          
#> x128         .          
#> x129         .          
#> x130         .          
#> x131         .          
#> x132         .          
#> x133         .          
#> x134         .          
#> x135         0.050432555
#> x136         .          
#> x137         .          
#> x138         .          
#> x139         .          
#> x140         .          
#> x141         .          
#> x142         .          
#> x143         .          
#> x144         .          
#> x145         .          
#> x146         .          
#> x147         .          
#> x148         .          
#> x149         .          
#> x150         .          
#> x151        -0.058158645
#> x152         .          
#> x153         .          
#> x154         .          
#> x155         .          
#> x156         .          
#> x157         .          
#> x158         .          
#> x159         0.089086626
#> x160         .          
#> x161         .          
#> x162         .          
#> x163         .          
#> x164         .          
#> x165         .          
#> x166         .          
#> x167         .          
#> x168        -0.288060753
#> x169         .          
#> x170         .          
#> x171         .          
#> x172         .          
#> x173         .          
#> x174         .          
#> x175         .          
#> x176         .          
#> x177         .          
#> x178         .          
#> x179         .          
#> x180         .          
#> x181         0.037796264
#> x182         .          
#> x183         0.042028658
#> x184         .          
#> x185         .          
#> x186         .          
#> x187         .          
#> x188         .          
#> x189         .          
#> x190         .          
#> x191         .          
#> x192         .          
#> x193         .          
#> x194         .          
#> x195         .          
#> x196         .          
#> x197         .          
#> x198         .          
#> x199         .          
#> x200         .          
#> x201         .          
#> x202        -0.019157694
#> x203         .          
#> x204         .          
#> x205         .          
#> x206         .          
#> x207         .          
#> x208         .          
#> x209         .          
#> x210         0.730390941
#> Degrees of Freedom: 139 Total (i.e. Null); 112 Residual
#> 
ypred <- predict.extendglmnet(fit.ridge, testx, family = "binomial", type="class")
evaluation(testy, ypred, "accuracy")
#> [1] 0.8833333

# random lasso
fit.rl <- extendedglmnet(trainx, trainy, family = "binomial", type = "random lasso", q1 = 150, q2 = 150)
print.extendglmnet(fit.rl)
#> 
#> Call: random lasso for binomial reponses
#> 
#>   (Intercept)            x1            x2            x3            x4 
#>  5.544756e-01  4.110122e-01  5.543854e-02  3.319326e-02 -9.210666e-02 
#>            x5            x6            x7            x8            x9 
#>  9.843695e-03  3.069360e-03 -2.082898e-01 -5.744327e-03  2.529208e-02 
#>           x10           x11           x12           x13           x14 
#>  0.000000e+00  1.725439e-03  0.000000e+00 -4.443563e-02 -1.679261e-02 
#>           x15           x16           x17           x18           x19 
#>  3.681454e-01  5.370547e-02  9.989181e-04  6.356109e-03  2.960483e-02 
#>           x20           x21           x22           x23           x24 
#>  5.860362e-03 -6.156394e-02  5.088243e-02 -2.864893e-04  9.436607e-04 
#>           x25           x26           x27           x28           x29 
#> -1.439721e-03 -6.524354e-02 -1.099171e-02 -1.588099e-01 -2.181786e-02 
#>           x30           x31           x32           x33           x34 
#>  7.883276e-02 -5.620761e-03  1.700197e-05 -2.183462e-03  3.894336e-03 
#>           x35           x36           x37           x38           x39 
#>  1.322280e-02 -1.039292e-01  8.835147e-02 -9.858975e-03 -6.126707e-04 
#>           x40           x41           x42           x43           x44 
#>  1.338632e+00 -3.698688e-02 -2.100170e-03  1.057580e-01 -2.463906e-04 
#>           x45           x46           x47           x48           x49 
#>  7.566692e-02 -1.125139e-01  1.821072e-04 -3.129029e-03  3.161387e-02 
#>           x50           x51           x52           x53           x54 
#> -2.233629e+00 -9.256536e-05 -5.399954e-02 -1.750332e-01 -3.334009e-02 
#>           x55           x56           x57           x58           x59 
#> -2.093917e-01  1.225706e-01 -5.192230e-03  3.168602e-02  3.342750e-02 
#>           x60           x61           x62           x63           x64 
#>  0.000000e+00 -1.144922e-02  1.009841e-01  1.935636e-01 -9.578526e-03 
#>           x65           x66           x67           x68           x69 
#> -3.706705e-02 -1.135373e-03 -1.005623e-01 -7.691371e-03  1.976882e-02 
#>           x70           x71           x72           x73           x74 
#>  2.095711e-02  0.000000e+00 -2.631305e-02  8.675853e-03 -2.125278e-03 
#>           x75           x76           x77           x78           x79 
#> -5.910903e-03 -2.626787e-02  1.141022e-03 -8.793201e-03 -9.796150e-02 
#>           x80           x81           x82           x83           x84 
#> -5.879015e-02  8.187199e-03 -1.258461e-03 -2.319206e-02  2.204335e-02 
#>           x85           x86           x87           x88           x89 
#> -7.832670e-03 -1.030324e-03 -1.615900e-01  1.457106e-02 -3.671547e-03 
#>           x90           x91           x92           x93           x94 
#>  9.455308e-04  3.814965e-02 -1.580508e-02  2.231642e-01  1.315085e-02 
#>           x95           x96           x97           x98           x99 
#>  0.000000e+00  7.746831e-03  0.000000e+00 -4.481472e-03  1.945035e-02 
#>          x100          x101          x102          x103          x104 
#> -1.427599e-01  3.153971e-01 -4.785760e-03 -3.834622e-03  1.489047e-01 
#>          x105          x106          x107          x108          x109 
#> -4.568787e-04 -4.590124e-02  3.514283e-03  2.507760e-02  2.492810e-02 
#>          x110          x111          x112          x113          x114 
#>  3.428978e-02 -8.219697e-02  9.619625e-03 -3.040070e-02 -8.274673e-03 
#>          x115          x116          x117          x118          x119 
#>  2.620027e-02  8.651182e-02  1.231395e-02  4.540642e-03 -4.025000e-03 
#>          x120          x121          x122          x123          x124 
#>  2.784477e-03  9.452089e-03 -2.655940e-01 -8.491245e-02  1.317751e-03 
#>          x125          x126          x127          x128          x129 
#> -1.124266e-02  1.780268e-02  1.923264e-02 -1.694571e-04 -6.242360e-03 
#>          x130          x131          x132          x133          x134 
#> -6.071222e-02  5.139203e-03 -2.143972e-02  0.000000e+00 -4.396376e-02 
#>          x135          x136          x137          x138          x139 
#>  1.496632e-01 -3.258242e-02 -5.074783e-04 -6.612300e-04 -2.849705e-03 
#>          x140          x141          x142          x143          x144 
#>  7.896741e-03  2.805498e-02  4.795288e-03  1.236056e-03  5.644679e-04 
#>          x145          x146          x147          x148          x149 
#>  1.024678e-04 -1.641783e-02 -4.621707e-03 -3.010610e-02  3.920158e-02 
#>          x150          x151          x152          x153          x154 
#>  4.522951e-02 -1.311422e-01 -1.311752e-02 -4.089690e-04  6.893016e-02 
#>          x155          x156          x157          x158          x159 
#> -9.329682e-02 -1.086513e-03  1.628833e-03  5.162079e-04  1.640487e-01 
#>          x160          x161          x162          x163          x164 
#> -2.652951e-02 -5.092640e-03  0.000000e+00 -3.029702e-03  2.977421e-02 
#>          x165          x166          x167          x168          x169 
#> -3.693680e-04  1.387549e-03 -5.117565e-03 -4.078270e-01  2.796707e-03 
#>          x170          x171          x172          x173          x174 
#> -9.965016e-03  1.724705e-03  0.000000e+00 -2.441325e-02 -5.701069e-03 
#>          x175          x176          x177          x178          x179 
#> -1.743703e-02  1.396995e-03 -3.909629e-04 -1.603641e-02 -1.145411e-03 
#>          x180          x181          x182          x183          x184 
#>  2.155058e-02  5.864646e-02  1.244710e-03  1.076275e-01 -5.413323e-04 
#>          x185          x186          x187          x188          x189 
#>  1.511673e-02  3.500970e-02  3.592874e-03 -2.192356e-02  9.269270e-03 
#>          x190          x191          x192          x193          x194 
#> -6.016184e-02 -1.411935e-03 -5.970129e-03 -8.972851e-02 -1.624829e-02 
#>          x195          x196          x197          x198          x199 
#>  1.125205e-02 -9.564562e-03 -3.294846e-03  8.011745e-03  5.383315e-02 
#>          x200          x201          x202          x203          x204 
#> -1.883142e-02 -8.581810e-04 -1.029463e-01 -1.095119e-01 -3.278403e-02 
#>          x205          x206          x207          x208          x209 
#>  1.108848e-01 -1.115722e-03  1.551559e-02  1.045245e-03 -5.583776e-03 
#>          x210 
#>  1.068933e+00 
#> Degrees of Freedom: 139 Total (i.e. Null); -62 Residual
#> 
ypred <- predict.extendglmnet(fit.rl, testx, family = "binomial", type="class")
evaluation(testy, ypred, "accuracy")
#> [1] 0.8833333

# continuous response example
size = 30
x = matrix(runif(35*size),ncol=35)
y = 10*x[,1] + 3*x[,2] + 20*(x[,3]-0.5)**2 + 10*x[,4] + 5*x[,35] + runif(1,0,1)

index = sample(1:size, 0.7*size)
trainx = x[index,]
testx = x[-index,]
trainy = y[index]
testy = y[-index]

fit.lr<- extendedglmnet(trainx, trainy, family = "gaussian", type = "regression")
#> 
#> Call:
#> lm(formula = y ~ ., data = dat)
#> 
#> Coefficients:
#> (Intercept)           x1           x2           x3           x4           x5  
#>     -88.429       28.108     -131.279      -14.321      118.820       50.698  
#>          x6           x7           x8           x9          x10          x11  
#>     -51.491        8.063       67.487      -18.141      111.016       20.066  
#>         x12          x13          x14          x15          x16          x17  
#>     -28.012      -10.820      -37.425        3.977       27.843       69.981  
#>         x18          x19          x20          x21          x22          x23  
#>      69.493      -26.034       -6.488           NA           NA           NA  
#>         x24          x25          x26          x27          x28          x29  
#>          NA           NA           NA           NA           NA           NA  
#>         x30          x31          x32          x33          x34          x35  
#>          NA           NA           NA           NA           NA           NA  
#> 
print.extendglmnet(fit.lr)
#> 
#> Call: linear regression for gaussian reponses
#> 
#> (Intercept)          x1          x2          x3          x4          x5 
#>  -88.428714   28.107665 -131.279049  -14.321161  118.820158   50.697979 
#>          x6          x7          x8          x9         x10         x11 
#>  -51.491393    8.062773   67.486923  -18.140680  111.015691   20.065643 
#>         x12         x13         x14         x15         x16         x17 
#>  -28.012056  -10.820304  -37.425158    3.977440   27.842860   69.980770 
#>         x18         x19         x20         x21         x22         x23 
#>   69.493308  -26.034096   -6.487872          NA          NA          NA 
#>         x24         x25         x26         x27         x28         x29 
#>          NA          NA          NA          NA          NA          NA 
#>         x30         x31         x32         x33         x34         x35 
#>          NA          NA          NA          NA          NA          NA 
#> Degrees of Freedom: 20 Total (i.e. Null); 0 Residual
#> 
ypred <- predict.extendglmnet(fit.lr, testx, family = "gaussian")
#> Warning: prediction from a rank-deficient fit may be misleading
evaluation(testy, ypred, "MSE")
#> [1] 4066.976

fit.lasso <- extendedglmnet(trainx, trainy, family = "gaussian", type = "lasso")
#> <sparse>[ <logic> ] : .M.sub.i.logical() maybe inefficient
print.extendglmnet(fit.lasso)
#> 
#> Call: lasso for gaussian reponses
#> 
#> 36 x 1 sparse Matrix of class "dgCMatrix"
#>                       s0
#> (Intercept)  3.181086628
#> x1           9.788270368
#> x2           2.090140052
#> x3           .          
#> x4           5.613719829
#> x5           .          
#> x6           .          
#> x7           .          
#> x8           .          
#> x9           .          
#> x10          .          
#> x11          .          
#> x12          .          
#> x13          .          
#> x14          .          
#> x15          .          
#> x16          0.002795373
#> x17          .          
#> x18          .          
#> x19          .          
#> x20          .          
#> x21          .          
#> x22          .          
#> x23          1.114928257
#> x24         -0.462166187
#> x25          .          
#> x26          0.291665554
#> x27          .          
#> x28          .          
#> x29          .          
#> x30          .          
#> x31          .          
#> x32          1.727231121
#> x33          .          
#> x34          .          
#> x35          4.943914903
#> Degrees of Freedom: 20 Total (i.e. Null); 11 Residual
#> 
ypred <- predict.extendglmnet(fit.lasso, testx, family = "gaussian")
evaluation(testy, ypred, "MSE")
#> [1] 7.387834

fit.ridge <- extendedglmnet(trainx, trainy, family = "gaussian", type = "ridge")
#> <sparse>[ <logic> ] : .M.sub.i.logical() maybe inefficient
print.extendglmnet(fit.ridge)
#> 
#> Call: ridge for gaussian reponses
#> 
#> 36 x 1 sparse Matrix of class "dgCMatrix"
#>                       s0
#> (Intercept) 12.127532992
#> x1           1.198593503
#> x2           0.879286934
#> x3           0.242887122
#> x4           1.365591905
#> x5           0.255758355
#> x6           0.197450624
#> x7           0.997269368
#> x8           0.747042689
#> x9           0.439283338
#> x10         -0.105220491
#> x11         -0.139801966
#> x12         -0.008720933
#> x13         -0.032937282
#> x14          0.241271541
#> x15          0.266335734
#> x16         -0.074296823
#> x17         -0.495506200
#> x18         -0.161325982
#> x19         -0.334371341
#> x20          0.024161216
#> x21         -0.322905167
#> x22          0.003546382
#> x23          0.483381058
#> x24         -0.402221835
#> x25         -0.035660501
#> x26          0.541713677
#> x27         -0.046932443
#> x28         -0.461040652
#> x29          0.129084711
#> x30         -0.269508668
#> x31         -0.333499267
#> x32          0.501120642
#> x33          0.345204773
#> x34         -0.352005158
#> x35          0.487673930
#> Degrees of Freedom: 20 Total (i.e. Null); -15 Residual
#> 
ypred <- predict.extendglmnet(fit.ridge, testx, family = "gaussian")
evaluation(testy, ypred, "MSE")
#> [1] 25.7101

fit.rl <- extendedglmnet(trainx, trainy, family = "gaussian", type = "random lasso")
print.extendglmnet(fit.rl)
#> 
#> Call: random lasso for gaussian reponses
#> 
#>  (Intercept)           x1           x2           x3           x4           x5 
#>  7.447301141  5.837996138  1.307712925 -0.015121444  5.034933046  0.096284018 
#>           x6           x7           x8           x9          x10          x11 
#>  0.000000000  1.890589643  0.662374422  0.138067140 -0.004859364  0.080595440 
#>          x12          x13          x14          x15          x16          x17 
#> -0.008377063 -0.040219338  0.016208401 -0.005783832  0.023961691 -0.133309870 
#>          x18          x19          x20          x21          x22          x23 
#> -0.217665215 -0.341191170 -0.006515488  0.016510661 -0.020260031  0.663869588 
#>          x24          x25          x26          x27          x28          x29 
#> -0.077878588 -0.071482135  0.483742273  0.152728560 -0.253296245  0.006488829 
#>          x30          x31          x32          x33          x34          x35 
#> -0.070991713 -0.224254675  0.305626710  0.089074822 -0.019832778  0.659570269 
#> Degrees of Freedom: 20 Total (i.e. Null); -14 Residual
#> 
ypred <- predict.extendglmnet(fit.rl, testx, family = "gaussian")
evaluation(testy, ypred, "MSE")
#> [1] 12.76297