make predictions from an extendedglmnet object.

Similar to other predict methods, this functions predicts fitted values and classes (for binary endpoint only) from a fitted "extendedglmnet" object.

# S3 method for predict
extendedglmnet(
  object,
  newx,
  type = c("link", "class", "response"),
  family = "gaussian",
  levels = c(0, 1)
)

Arguments

newx

Matrix of new values for x at which predictions are to be made. Must be a matrix.

type

Type of prediction required. For continuous y's, only predicted responses will be returned. Type "link" gives the linear predictors for "binomial" models. Type "response" gives the fitted probabilities for "binomial" models. Type "class" applies only to "binomial" models, and produces the class label corresponding to the maximum probability.

family

can be binomial or gaussian depending on y types.

levels

specify the binary response variable levels of factors.

fit

Fitted "extendedglmnet" model object.

Value

fitted values.

See also

Author

Weiwei Tao tww101928@gmail.com

Examples

size = 200
x = matrix(rnorm(210*size),ncol=210)
z = 1 + 2*x[,1] + 3*x[,40] - 4*x[,50] + 3*x[,210]
pr = 1/(1+exp(-z))
y = rbinom(size,1,pr)

fit.rl <- extendedglmnet(x, y, family = "binomial", type = "random lasso", q1 = 150, q2 = 150)
ypred.rl <- predict.extendglmnet(fit.rl, x, type="class", family = "binomial")