print an extendedglmnet fitting results

print an extendedglmnet fitting results

# S3 method for print
extendedglmnet(x)

Arguments

fit

fitted extendedglmnet object

Details

The call that the fitting results from extendedglmnet is printed. The coefficients of the fit followed by degrees of freedom (nobs - 1) and the residual degree of freedom (nobs - number of nonzero coefficients).

See also

Author

Weiwei Tao tww101928@gmail.com

Examples

size = 30
x = matrix(runif(35*size),ncol=35)
y = 10*x[,1] + 3*x[,2] + 20*(x[,3]-0.5)**2 + 10*x[,4] + 5*x[,35] + runif(1,0,1)
fit.regression <- extendedglmnet(x, y, family = "gaussian", type = "regression")
#> 
#> Call:
#> lm(formula = y ~ ., data = dat)
#> 
#> Coefficients:
#> (Intercept)           x1           x2           x3           x4           x5  
#>    11.57511      4.96908      4.35327     -0.09445     -5.41604      2.20312  
#>          x6           x7           x8           x9          x10          x11  
#>    13.43423     -9.91574      3.95734      9.27023     -6.99583     11.61067  
#>         x12          x13          x14          x15          x16          x17  
#>     6.86252      5.97403     -3.35530     -2.71821     -6.16120      2.27768  
#>         x18          x19          x20          x21          x22          x23  
#>     5.85425      2.90071    -29.90491     -0.29176      3.77263     22.32550  
#>         x24          x25          x26          x27          x28          x29  
#>     2.28661     -0.58917      3.77521      1.24923      0.86181    -24.55991  
#>         x30          x31          x32          x33          x34          x35  
#>          NA           NA           NA           NA           NA           NA  
#> 
print.extendglmnet(fit.regression)
#> 
#> Call: linear regression for gaussian reponses
#> 
#> (Intercept)          x1          x2          x3          x4          x5 
#>  11.5751113   4.9690771   4.3532744  -0.0944506  -5.4160361   2.2031229 
#>          x6          x7          x8          x9         x10         x11 
#>  13.4342349  -9.9157366   3.9573393   9.2702327  -6.9958300  11.6106707 
#>         x12         x13         x14         x15         x16         x17 
#>   6.8625209   5.9740274  -3.3552977  -2.7182091  -6.1612030   2.2776816 
#>         x18         x19         x20         x21         x22         x23 
#>   5.8542523   2.9007146 -29.9049142  -0.2917589   3.7726338  22.3255028 
#>         x24         x25         x26         x27         x28         x29 
#>   2.2866130  -0.5891706   3.7752110   1.2492338   0.8618118 -24.5599079 
#>         x30         x31         x32         x33         x34         x35 
#>          NA          NA          NA          NA          NA          NA 
#> Degrees of Freedom: 29 Total (i.e. Null); 0 Residual
#>